flask-server/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
supabase
PyPDF2
jina
pypdf
fastapi
uvicorn
python-multipart